backend/requirements.txt
# Backend Dependencies (Python 3.13 compatible)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
pgvector>=0.2.4

# AI Integration
openai>=1.10.0
google-generativeai>=0.3.2

# PDF Generation
reportlab>=4.0.9

# Utilities
python-multipart>=0.0.6
httpx>=0.26.0
